+++ tests/test_course_admin/test_course_permissions/test_handgrader_roster.ts
@@ -1,5 +1,5 @@
-import HandgraderRoster from '@/components/course_admin/permissions/handgrader_roster.vue';
-import Permissions from '@/components/course_admin/permissions/permissions.vue';
+import HandgraderRoster from '@/components/course_admin/roster/handgrader_roster.vue';
+import Permissions from '@/components/course_admin/roster/permissions.vue';
 import { config, mount, Wrapper } from '@vue/test-utils';
 import { Course, Semester, User } from 'ag-client-typescript';
