+++ tests/test_course_admin/test_course_permissions/test_staff_roster.ts
@@ -1,5 +1,5 @@
-import Permissions from '@/components/course_admin/permissions/permissions.vue';
-import StaffRoster from '@/components/course_admin/permissions/staff_roster.vue';
+import Permissions from '@/components/course_admin/roster/permissions.vue';
+import StaffRoster from '@/components/course_admin/roster/staff_roster.vue';
 import { config, mount, Wrapper } from '@vue/test-utils';
 import { Course, Semester, User } from 'ag-client-typescript';
